docker/backend/src/email_service/requirements.txt
azure-communication-email
azure-identity
